+++ packages/manager/apps/container/src/container/nav-reshuffle/header/user-account-menu/constants.ts
@@ -18,7 +18,7 @@ export const tracking = {
 
 export const links: UserLink[] = [
   {
-    app: 'new-account',
+    app: 'account',
     key: 'user-account-menu-profile',
     hash: '#/useraccount/dashboard',
     i18nKey: 'user_account_menu_profile',
@@ -39,11 +39,20 @@ export const links: UserLink[] = [
     trackingHit: tracking.myServices,
   },
   {
-    app: 'new-account',
+    app: 'new-billing',
+    key: 'myContracts',
+    hash: '#/autorenew/agreements',
+    i18nKey: 'user_account_menu_my_contracts',
+    trackingHit: tracking.contracts,
+    region: ['EU', 'CA'],
+  },
+  {
+    app: 'account',
     key: 'myCommunications',
     hash: '#/useraccount/emails',
     i18nKey: 'user_account_menu_my_communication',
     trackingHit: tracking.accountContacts,
+    region: ['EU', 'CA'],
   },
   {
     app: 'new-billing',
@@ -60,10 +69,18 @@ export const links: UserLink[] = [
     trackingHit: tracking.orders,
   },
   {
-    app: 'new-account',
+    app: 'account',
     key: 'myContacts',
     hash: '#/contacts/services',
     i18nKey: 'user_account_menu_my_contacts',
     trackingHit: tracking.contacts,
+    region: ['EU', 'CA'],
   },
+  {
+    app: 'dedicated',
+    key: 'myAssistanceTickets',
+    hash: '#/ticket',
+    i18nKey: 'user_account_menu_my_assistance_tickets',
+    region: ['US'],
+  }
 ];