+++ packages/manager/apps/pci-billing/src/api/hook/useConsumption.tsx
@@ -147,6 +147,40 @@ const initSnapshotList = (data: TCurrentUsage) => {
   };
 };
 
+const initRancherList = (
+  data: TCurrentUsage,
+): {
+  rancherList: TResourceUsage[];
+  rancherTotalPrice: number;
+} => {
+  if (!data.hourlyUsage.rancher.length) {
+    return {
+      rancherList: [],
+      rancherTotalPrice: 0,
+    };
+  }
+
+  const rancherList = data.hourlyUsage.rancher.flatMap((rancher) =>
+    rancher.details.map((r) => ({
+      ...r,
+      totalPrice: roundPrice(r.totalPrice.value),
+      name: rancher.reference,
+      region: rancher.region,
+    })),
+  );
+
+  const rancherTotalPrice = roundPrice(
+    data.hourlyUsage.rancher
+      .flatMap((rancher) => rancher.details)
+      .reduce((sum, rancher) => sum + roundPrice(rancher.totalPrice.value), 0),
+  );
+
+  return {
+    rancherList,
+    rancherTotalPrice,
+  };
+};
+
 const initVolumeList = (data: TCurrentUsage) => {
   if (!data?.hourlyUsage) {
     return {
@@ -361,8 +395,10 @@ export type TConsumptionDetail = {
   volumes: TVolume[];
   bandwidthByRegions: TInstanceBandWith[];
   privateRegistry: TResourceUsage[];
+  rancher: TResourceUsage[];
   kubernetesLoadBalancer: TResourceUsage[];
   training: TResourceUsage[];
+  aiEndpoints: TResourceUsage[];
   notebooks: TResourceUsage[];
   aiDeploy: TResourceUsage[];
   coldArchive: TResourceUsage[];
@@ -381,6 +417,7 @@ export type TConsumptionDetail = {
       archiveStorage: number;
       snapshot: number;
       volume: number;
+      rancher: number;
       bandwidth: number;
       privateRegistry: number;
       kubernetesLoadBalancer: number;
@@ -388,6 +425,7 @@ export type TConsumptionDetail = {
       coldArchive: number;
       serving: number;
       training: number;
+      aiEndpoints: number;
       aiDeploy: number;
       dataProcessing: number;
       databases: number;
@@ -412,10 +450,12 @@ export const initializeTConsumptionDetail = (): TConsumptionDetail => ({
   volumes: [],
   bandwidthByRegions: [],
   privateRegistry: [],
+  rancher: [],
   kubernetesLoadBalancer: [],
   training: [],
   notebooks: [],
   aiDeploy: [],
+  aiEndpoints: [],
   coldArchive: [],
   dataProcessing: [],
   databases: [],
@@ -432,6 +472,7 @@ export const initializeTConsumptionDetail = (): TConsumptionDetail => ({
       archiveStorage: 0,
       snapshot: 0,
       volume: 0,
+      rancher: 0,
       bandwidth: 0,
       privateRegistry: 0,
       kubernetesLoadBalancer: 0,
@@ -440,6 +481,7 @@ export const initializeTConsumptionDetail = (): TConsumptionDetail => ({
       serving: 0,
       training: 0,
       aiDeploy: 0,
+      aiEndpoints: 0,
       dataProcessing: 0,
       databases: 0,
       floatingIP: 0,
@@ -461,6 +503,7 @@ export const getConsumptionDetails = (usage: TCurrentUsage) => {
     { type: 'ai-notebook', key: 'notebooks' },
     { type: 'ai-serving-engine', key: 'serving' },
     { type: 'ai-training', key: 'training' },
+    { type: 'ai-endpoints', key: 'aiEndpoints' },
     { type: 'data-processing-job', key: 'dataProcessing' },
     { type: 'databases', key: 'databases' },
     { type: 'coldarchive', key: 'coldArchive' },
@@ -504,7 +547,7 @@ export const getConsumptionDetails = (usage: TCurrentUsage) => {
   const { snapshotList, snapshotsTotalPrice } = initSnapshotList(usage);
   const { volumeList, volumesTotalPrice } = initVolumeList(usage);
   const { bandwidthList, bandwidthTotalPrice } = initInstanceBandwidth(usage);
-
+  const { rancherList, rancherTotalPrice } = initRancherList(usage);
   const totals = {
     hourly: {
       total: 0,
@@ -515,6 +558,7 @@ export const getConsumptionDetails = (usage: TCurrentUsage) => {
       snapshot: snapshotsTotalPrice,
       volume: volumesTotalPrice,
       bandwidth: bandwidthTotalPrice,
+      rancher: rancherTotalPrice,
     },
     monthly: {
       total: monthlyInstanceTotalPrice,
@@ -534,6 +578,7 @@ export const getConsumptionDetails = (usage: TCurrentUsage) => {
     objectStorages: objectStorageList,
     archiveStorages: archiveStorageList,
     snapshots: snapshotList,
+    rancher: rancherList,
     volumes: volumeList,
     bandwidthByRegions: bandwidthList,
     ...resources,